requirements.txt
pyrtlsdr
scipy
numpy
si_prefix